Isolate serial transport layer and refactor remote server
