+++ picobot/__main__.py
@@ -1,0 +1,7 @@
+"""Run the PicoBot GUI via ``python -m picobot``."""
+
+from .app import main
+
+
+if __name__ == "__main__":
+    main()
