+++ picobot/playback/__init__.py
@@ -1,0 +1,1 @@
+"""Macro playback orchestration for PicoBot."""
