+++ picobot/transport/__init__.py
@@ -1,0 +1,5 @@
+"""Transport layer abstractions for PicoBot."""
+
+from .serial_manager import SerialManager, SerialSession, SerialTransportError
+
+__all__ = ["SerialManager", "SerialSession", "SerialTransportError"]
