+++ picobot/ui/__init__.py
@@ -1,0 +1,1 @@
+"""User interface components for PicoBot."""
